Running OpenSCAP Scans with MCollective Slide Deck
This is my presentation from the 2015 Puppet Camp DC.

The premise is to have both scan discovery capabilities as well as
targeted scan capabilities across your entire fleet of connected
systems.

Also...I just wanted to see if it would work!

The associated code can be found in the Onyx Point Git repo.